refactor to pull model builder into funtions
