+++ cypress/e2e/internal/return-logs/historic-corrections-03.cy.js
@@ -1,22 +1,8 @@
 'use strict'
 
-import licence from '../../../support/fixture-builder/licence.js'
-import points from '../../../support/fixture-builder/points.js'
-import purposes from '../../../support/fixture-builder/purposes.js'
-import returnLogs from '../../../support/fixture-builder/return-logs.js'
-import returnRequirements from '../../../support/fixture-builder/return-requirements.js'
-import returnRequirementPoints from '../../../support/fixture-builder/return-requirement-points.js'
-import returnVersion from '../../../support/fixture-builder/return-version.js'
+import { basicLicenceTwoReturnRequirementsWithThreeReturnLogsEach } from '../../../support/fixture-builder/scenarios.js'
 
-const dataModel = {
-  ...licence(),
-  ...points(2),
-  ...purposes(2),
-  ...returnVersion(),
-  ...returnRequirements(2),
-  ...returnRequirementPoints(2),
-  ...returnLogs(7)
-}
+const dataModel = basicLicenceTwoReturnRequirementsWithThreeReturnLogsEach()
 
 describe('Submit historic correction using abstraction data for two abstraction points', () => {
   beforeEach(() => {
@@ -34,68 +20,6 @@ describe('Submit historic correction using abstraction data for two abstraction 
     // day and month we'll get back 20XX-03-31.
     cy.currentFinancialYear().then((currentFinancialYearInfo) => {
       cy.wrap(currentFinancialYearInfo).as('currentFinancialYearInfo')
-
-      dataModel.returnRequirements[1].summer = false
-
-      let startYear = currentFinancialYearInfo.start.year
-      let endYear = currentFinancialYearInfo.end.year
-
-      dataModel.returnLogs[0].id = `v1:9:AT/CURR/DAILY/01:9999990:${startYear}-04-01:${endYear}-03-31`
-      dataModel.returnLogs[0].dueDate = `${endYear}-04-28`
-      dataModel.returnLogs[0].endDate = `${endYear}-03-31`
-      dataModel.returnLogs[0].metadata.isSummer = false
-      dataModel.returnLogs[0].startDate = `${startYear}-04-01`
-      dataModel.returnLogs[0].status = 'due'
-      dataModel.returnLogs[0].returnCycleId.value = `${startYear}-04-01`
-
-      dataModel.returnLogs[1].id = `v1:9:AT/CURR/DAILY/01:9999991:${startYear}-04-01:${endYear}-03-31`
-      dataModel.returnLogs[1].dueDate = `${endYear}-04-28`
-      dataModel.returnLogs[1].endDate = `${endYear}-03-31`
-      dataModel.returnLogs[1].metadata.isSummer = false
-      dataModel.returnLogs[1].startDate = `${startYear}-04-01`
-      dataModel.returnLogs[1].status = 'due'
-      dataModel.returnLogs[1].returnCycleId.value = `${startYear}-04-01`
-      dataModel.returnLogs[1].returnReference = '9999991'
-
-      startYear = startYear - 1
-      endYear = endYear - 1
-
-      dataModel.returnLogs[2].id = `v1:9:AT/CURR/DAILY/01:9999990:${startYear}-04-01:${endYear}-03-31`
-      dataModel.returnLogs[2].dueDate = `${endYear}-04-28`
-      dataModel.returnLogs[2].endDate = `${endYear}-03-31`
-      dataModel.returnLogs[2].metadata.isSummer = false
-      dataModel.returnLogs[2].startDate = `${startYear}-04-01`
-      dataModel.returnLogs[2].status = 'completed'
-      dataModel.returnLogs[2].returnCycleId.value = `${startYear}-04-01`
-
-      dataModel.returnLogs[3].id = `v1:9:AT/CURR/DAILY/01:9999991:${startYear}-04-01:${endYear}-03-31`
-      dataModel.returnLogs[3].dueDate = `${endYear}-04-28`
-      dataModel.returnLogs[3].endDate = `${endYear}-03-31`
-      dataModel.returnLogs[3].metadata.isSummer = false
-      dataModel.returnLogs[3].startDate = `${startYear}-04-01`
-      dataModel.returnLogs[3].status = 'completed'
-      dataModel.returnLogs[3].returnCycleId.value = `${startYear}-04-01`
-      dataModel.returnLogs[3].returnReference = '9999991'
-
-      startYear = startYear - 1
-      endYear = endYear - 1
-
-      dataModel.returnLogs[4].id = `v1:9:AT/CURR/DAILY/01:9999990:${startYear}-04-01:${endYear}-03-31`
-      dataModel.returnLogs[4].dueDate = `${endYear}-04-28`
-      dataModel.returnLogs[4].endDate = `${endYear}-03-31`
-      dataModel.returnLogs[4].metadata.isSummer = false
-      dataModel.returnLogs[4].startDate = `${startYear}-04-01`
-      dataModel.returnLogs[4].status = 'completed'
-      dataModel.returnLogs[4].returnCycleId.value = `${startYear}-04-01`
-
-      dataModel.returnLogs[5].id = `v1:9:AT/CURR/DAILY/01:9999991:${startYear}-04-01:${endYear}-03-31`
-      dataModel.returnLogs[5].dueDate = `${endYear}-04-28`
-      dataModel.returnLogs[5].endDate = `${endYear}-03-31`
-      dataModel.returnLogs[5].metadata.isSummer = false
-      dataModel.returnLogs[5].startDate = `${startYear}-04-01`
-      dataModel.returnLogs[5].status = 'completed'
-      dataModel.returnLogs[5].returnCycleId.value = `${startYear}-04-01`
-      dataModel.returnLogs[5].returnReference = '9999991'
 
       cy.load(dataModel)
     })
@@ -216,22 +140,19 @@ describe('Submit historic correction using abstraction data for two abstraction 
 
       cy.returnLogDueData(startYear - 1, true).then((data) => {
         cy.get('[data-test="return-due-date-8"]').contains(data.text)
-        cy.get('[data-test="return-status-8"] > .govuk-tag').contains('void')
+        cy.get('[data-test="return-status-8"] > .govuk-tag').contains(data.label)
 
         cy.get('[data-test="return-due-date-9"]').contains(data.text)
-        cy.get('[data-test="return-status-9"] > .govuk-tag').contains('void')
+        cy.get('[data-test="return-status-9"] > .govuk-tag').contains(data.label)
 
         cy.get('[data-test="return-due-date-10"]').contains(data.text)
-        cy.get('[data-test="return-status-10"] > .govuk-tag').contains(data.label)
+        cy.get('[data-test="return-status-10"] > .govuk-tag').contains('void')
 
         cy.get('[data-test="return-due-date-11"]').contains(data.text)
-        cy.get('[data-test="return-status-11"] > .govuk-tag').contains(data.label)
+        cy.get('[data-test="return-status-11"] > .govuk-tag').contains('void')
 
         cy.get('[data-test="return-due-date-12"]').contains(data.text)
         cy.get('[data-test="return-status-12"] > .govuk-tag').contains(data.label)
-
-        cy.get('[data-test="return-due-date-13"]').contains(data.text)
-        cy.get('[data-test="return-status-13"] > .govuk-tag').contains(data.label)
       })
     })
   })
