+++ cypress/e2e/internal/return-logs/quarterly-01.cy.js
@@ -1,22 +1,8 @@
 'use strict'
 
-import licence from '../../../support/fixture-builder/licence.js'
-import points from '../../../support/fixture-builder/points.js'
-import purposes from '../../../support/fixture-builder/purposes.js'
-import returnLogs from '../../../support/fixture-builder/return-logs.js'
-import returnRequirements from '../../../support/fixture-builder/return-requirements.js'
-import returnRequirementPoints from '../../../support/fixture-builder/return-requirement-points.js'
-import returnVersion from '../../../support/fixture-builder/return-version.js'
+import { basicLicenceOneReturnRequirementQuarterlyReturns } from '../../../support/fixture-builder/scenarios.js'
 
-const dataModel = {
-  ...licence(),
-  ...points(),
-  ...purposes(),
-  ...returnVersion(),
-  ...returnRequirements(),
-  ...returnRequirementPoints(),
-  ...returnLogs(4)
-}
+const dataModel = basicLicenceOneReturnRequirementQuarterlyReturns()
 
 describe('Submit winter and all year quarterly historic correction using abstraction data', () => {
   let year = new Date().getFullYear()
@@ -35,30 +21,6 @@ describe('Submit winter and all year quarterly historic correction using abstrac
       })
     })
 
-    dataModel.returnLogs[0].id = `v1:9:AT/CURR/DAILY/01:9999990:${year + 1}-01-01:${year + 1}-03-31`
-    dataModel.returnLogs[0].dueDate = `${year + 1}-04-28`
-    dataModel.returnLogs[0].endDate = `${year + 1}-03-31`
-    dataModel.returnLogs[0].startDate = `${year + 1}-01-01`
-    dataModel.returnLogs[0].returnCycleId.value = `${year}-04-01`
-
-    dataModel.returnLogs[1].id = `v1:9:AT/CURR/DAILY/01:9999990:${year}-10-01:${year}-12-31`
-    dataModel.returnLogs[1].dueDate = `${year + 1}-01-28`
-    dataModel.returnLogs[1].endDate = `${year}-12-31`
-    dataModel.returnLogs[1].startDate = `${year}-10-01`
-    dataModel.returnLogs[1].returnCycleId.value = `${year}-04-01`
-
-    dataModel.returnLogs[2].id = `v1:9:AT/CURR/DAILY/01:9999990:${year}-07-01:${year}-09-30`
-    dataModel.returnLogs[2].dueDate = `${year}-10-28`
-    dataModel.returnLogs[2].endDate = `${year}-09-30`
-    dataModel.returnLogs[2].startDate = `${year}-07-01`
-    dataModel.returnLogs[2].returnCycleId.value = `${year}-04-01`
-
-    dataModel.returnLogs[3].id = `v1:9:AT/CURR/DAILY/01:9999990:${year}-04-01:${year}-06-30`
-    dataModel.returnLogs[3].dueDate = `${year}-07-28`
-    dataModel.returnLogs[3].endDate = `${year}-06-30`
-    dataModel.returnLogs[3].startDate = `${year}-04-01`
-    dataModel.returnLogs[3].returnCycleId.value = `${year}-04-01`
-
     cy.load(dataModel)
   })
 
@@ -67,14 +29,26 @@ describe('Submit winter and all year quarterly historic correction using abstrac
 
     // confirm we are on the licence returns tab and that there are previous reuturn logs
     cy.get('#returns > .govuk-heading-l').contains('Returns')
-    cy.get('[data-test="return-due-date-0"]').contains(`28 April ${year + 1}`)
-    cy.get('[data-test="return-status-0"] > .govuk-tag').contains('not due yet')
-    cy.get('[data-test="return-due-date-1"]').contains(`28 January ${year + 1}`)
-    cy.get('[data-test="return-status-1"] > .govuk-tag').contains('complete')
-    cy.get('[data-test="return-due-date-2"]').contains(`28 October ${year}`)
-    cy.get('[data-test="return-status-2"] > .govuk-tag').contains('complete')
-    cy.get('[data-test="return-due-date-3"]').contains(`28 July ${year}`)
-    cy.get('[data-test="return-status-3"] > .govuk-tag').contains('complete')
+
+    cy.quarterlyReturnLogDueData(`${year + 1}-04-28`).then((data) => {
+      cy.get('[data-test="return-due-date-0"]').contains(data.text)
+      cy.get('[data-test="return-status-0"] > .govuk-tag').contains(data.label)
+    })
+
+    cy.quarterlyReturnLogDueData(`${year + 1}-01-28`).then((data) => {
+      cy.get('[data-test="return-due-date-1"]').contains(data.text)
+      cy.get('[data-test="return-status-1"] > .govuk-tag').contains('complete')
+    })
+
+    cy.quarterlyReturnLogDueData(`${year}-10-28`).then((data) => {
+      cy.get('[data-test="return-due-date-2"]').contains(data.text)
+      cy.get('[data-test="return-status-2"] > .govuk-tag').contains('complete')
+    })
+
+    cy.quarterlyReturnLogDueData(`${year}-07-28`).then((data) => {
+      cy.get('[data-test="return-due-date-3"]').contains(data.text)
+      cy.get('[data-test="return-status-3"] > .govuk-tag').contains('complete')
+    })
 
     // click licence set up tab
     cy.contains('Licence set up').click()
