+++ cypress/e2e/internal/return-versions/copy-existing.cy.js
@@ -1,20 +1,8 @@
 'use strict'
 
-import licence from '../../../support/fixture-builder/licence.js'
-import points from '../../../support/fixture-builder/points.js'
-import purposes from '../../../support/fixture-builder/purposes.js'
-import returnRequirements from '../../../support/fixture-builder/return-requirements.js'
-import returnRequirementPoints from '../../../support/fixture-builder/return-requirement-points.js'
-import returnVersion from '../../../support/fixture-builder/return-version.js'
+import { basicLicenceOneReturnRequirementsWithTwoPoints } from '../../../support/fixture-builder/scenarios.js'
 
-const dataModel = {
-  ...licence(),
-  ...points(2),
-  ...purposes(2),
-  ...returnVersion(),
-  ...returnRequirements(),
-  ...returnRequirementPoints(2)
-}
+const dataModel = basicLicenceOneReturnRequirementsWithTwoPoints()
 
 describe('Submit returns requirement using copy existing (internal)', () => {
   beforeEach(() => {
@@ -27,8 +15,6 @@ describe('Submit returns requirement using copy existing (internal)', () => {
         email: userEmail
       })
     })
-
-    dataModel.returnRequirementPoints[1].returnRequirementId = dataModel.returnRequirements[0].id
 
     cy.load(dataModel)
   })
